requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14,<4.0
python-dotenv>=1.0
